Get rid of some platform-specific implementations
